Add missing metadata_expire and cost settings to repository definition
Add cost to repository definitions, to make it deterministic which one
is chosen. Otherwise when the same package is in multiple repositories
(like stable and testing) and is downloaded using --downloadonly option,
dnf may try to re-download it from the other repository. The operation
would intentionally fail (thanks to setting invalid proxy), so the
template build.

Also add explicit metadata_expire.